adds analyst helper folder to hold repo specific analysis tools
