+++ lib/analyst_helper/helper.rb
@@ -1,0 +1,6 @@
+require_relative './item_analyst_helper'
+
+# holds references to analyst helper modules
+module AnalystHelper
+  include ItemAnalyst
+end
